# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "contact": {
      "name": "Ferme des animaux - IIM",
      "url": "https://github.com/alyssialopr/FermeDesAnimauxJava"
    },
    "description": "Gestion d'une ferme : des **eleveurs** achetent, vendent, nourrissent, soignent et promenent des **vaches** et des **poules**.\n\n### Principes\n\n- un animal peut arriver a la ferme sans proprietaire, puis etre achete ;\n- un eleveur ne peut agir que sur ses propres animaux, sinon la reponse est `409` avec le message *Cet animal ne vous appartient pas* ;\n- un animal `VENDU`, `MORT` ou `DISPARU` n'accepte plus aucune action ;\n- chaque action renvoie le message metier correspondant ainsi que l'etat de l'animal apres l'action.\n\n### Erreurs\n\nToutes les erreurs suivent la RFC 7807 (`application/problem+json`) : `title`, `status`, `detail`, `instance`, `horodatage`, et `champs` pour les erreurs de validation.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "API Ferme des animaux",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Environnement local (docker compose)"
    }
  ],
...
  "paths": {
    "/api/animaux": {
      "get": {
        "description": "Les filtres se combinent. Sans filtre, tout le cheptel est renvoye, trie par nom.",
        "operationId": "recupereAnimaux",
        "parameters": [
          {
            "description": "Ne garder qu'une espece",
            "example": "POULE",
            "in": "query",
            "name": "espece",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Espece de l'animal. Elle determine sa production, ses tarifs et la vitesse a laquelle il a faim. Le catalogue complet est expose par `GET /api/especes`.",
              "enum": [
                "VACHE",
                "CHEVRE",
                "MOUTON",
                "POULE",
                "CANARD",
                "OIE",
                "LAPIN",
                "COCHON",
                "CHEVAL"
              ]
            }
          },
          {
            "description": "Ne garder que les animaux dans cet etat",
            "example": "LIBRE",
            "in": "query",
            "name": "etat",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Etat de l'animal. `LIBRE` : present a la ferme, toutes les actions sont possibles. `VENDU` : sorti du troupeau apres une vente. `MORT` et `DISPARU` : sorties definitives declarees par la ferme. Hors `LIBRE`, aucune action n'est acceptee.",
              "enum": [
                "LIBRE",
                "VENDU",
                "MORT",
                "DISPARU"
              ]
            }
          },
          {
            "description": "Ne garder que le troupeau de cet eleveur",
            "example": 1,
            "in": "query",
            "name": "eleveurId",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "description": "Ne garder que les animaux de cet enclos",
            "example": 3,
            "in": "query",
            "name": "enclos",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnimalResponse"
                  }
                }
              }
            },
            "description": "Cheptel correspondant aux filtres"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Lister les animaux, avec filtres optionnels",
        "tags": [
          "Animaux"
        ]
      },
      "post": {
        "description": "Cree une vache ou une poule. Les champs specifiques a l'espece (`litresDeLaitParJour`, `oeufsParSemaine`) sont facultatifs et prennent une valeur par defaut. Si `eleveurId` est fourni, l'animal est achete dans la foulee par cet eleveur.",
        "operationId": "creerAnimal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerAnimalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            },
            "description": "Animal enregistre, son adresse est dans l'en-tete Location"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Faire entrer un animal a la ferme",
        "tags": [
          "Animaux"
        ]
      }
    },
    "/api/animaux/{id}": {
      "delete": {
        "description": "Efface l'animal de la base. Pour tracer une sortie, preferer PATCH /etat.",
        "operationId": "supprimerAnimal",
        "parameters": [
          {
            "description": "Identifiant de l'animal",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Animal supprime"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Retirer definitivement un animal du registre",
        "tags": [
          "Animaux"
        ]
      },
      "get": {
        "operationId": "recupereAnimalParId",
        "parameters": [
          {
            "description": "Identifiant de l'animal",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            },
            "description": "Fiche de l'animal"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Recuperer un animal par son identifiant",
        "tags": [
          "Animaux"
        ]
      }
    },
    "/api/animaux/{id}/enclos": {
      "patch": {
        "description": "Refuse si l'animal n'est plus a la ferme (vendu, mort ou disparu).",
        "operationId": "changerEnclos",
        "parameters": [
          {
            "description": "Identifiant de l'animal",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MajEnclosRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            },
            "description": "Animal deplace"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Deplacer un animal dans un autre enclos",
        "tags": [
          "Animaux"
        ]
      }
    },
    "/api/animaux/{id}/etat": {
      "patch": {
        "description": "Seuls `MORT` et `DISPARU` sont acceptes ici : la vente passe par l'eleveur et un animal deja sorti du cheptel ne peut plus changer d'etat.",
        "operationId": "changerEtat",
        "parameters": [
          {
            "description": "Identifiant de l'animal",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MajEtatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalResponse"
                }
              }
            },
            "description": "Etat mis a jour"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Validation": {
                    "description": "Validation",
                    "value": {
                      "title": "Requete invalide",
                      "status": 400,
                      "detail": "Certains champs sont invalides.",
                      "instance": "/api/animaux",
                      "horodatage": "2026-09-18T07:27:13.637Z",
                      "champs": {
                        "nom": "le nom est obligatoire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Requete invalide : champ manquant, valeur d'enumeration inconnue ou corps illisible"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Introuvable": {
                    "description": "Introuvable",
                    "value": {
                      "title": "Ressource introuvable",
                      "status": 404,
                      "detail": "Animal introuvable : 42",
                      "instance": "/api/animaux/42",
                      "horodatage": "2026-09-18T07:27:13.544Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Animal ou eleveur inconnu"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "Action impossible": {
                    "description": "Action impossible",
                    "value": {
                      "title": "Action impossible",
                      "status": 409,
                      "detail": "Impossible de nourrir la vache emily : l'animal est VENDU.",
                      "instance": "/api/animaux/2/etat",
                      "horodatage": "2026-09-18T07:27:13.445Z"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            },
            "description": "Regle metier : animal vendu, mort ou disparu"
          }
        },
        "summary": "Declarer un animal mort ou disparu",
        "tags": [
          "Animaux"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Erreur": {
        "type": "object",
        "description": "Erreur normalisee au format RFC 7807 (application/problem+json)",
        "properties": {
          "champs": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Detail par champ, uniquement pour les erreurs de validation (400)",
            "example": {
              "nom": "le nom est obligatoire"
            }
          },
          "detail": {
            "type": "string",
            "description": "Message explicatif",
            "example": "Cet animal ne vous appartient pas"
          },
          "horodatage": {
            "type": "string",
            "format": "date-time",
            "description": "Date de l'erreur",
            "example": "2026-09-18T07:27:13.445Z"
          },
          "instance": {
            "type": "string",
            "description": "Chemin de la requete fautive",
            "example": "/api/eleveurs/2/animaux/1/repas"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "Code HTTP",
            "example": 409
          },
          "title": {
            "type": "string",
            "description": "Intitule court de l'erreur",
            "example": "Animal non possede"
          }
        }
      },
...
      "MajEnclosRequest": {
        "type": "object",
        "description": "Nouvel enclos de l'animal",
        "properties": {
          "enclos": {
            "type": "string",
            "description": "Enclos de destination",
            "example": "7",
            "maxLength": 40,
            "minLength": 0,
            "pattern": "[\\p{L}\\p{N} '\\-]+"
          }
        },
        "required": [
          "enclos"
        ]
      },
      "MajEtatRequest": {
        "type": "object",
        "description": "Declaration de sortie definitive d'un animal",
        "properties": {
          "etat": {
            "type": "string",
            "description": "Nouvel etat, MORT ou DISPARU uniquement. La vente se fait via l'eleveur et un animal ne revient jamais a l'etat LIBRE par cette route",
            "enum": [
              "MORT",
              "DISPARU"
            ],
            "example": "DISPARU"
          }
        },
        "required": [
          "etat"
        ]
      },
...
    }
  }
}